refactor(platform): extract deriveRosterItems, fix sidebar badge consistency
